+++ modules/prebidServerBidAdapter/config.js
