+++ test/spec/modules/prebidServerBidAdapter_spec.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { PrebidServer as Adapter, resetSyncedStatus } from 'modules/prebidServerBidAdapter';
+import { PrebidServer as Adapter, resetSyncedStatus } from 'modules/prebidServerBidAdapter/index.js';
 import adapterManager from 'src/adaptermanager';
 import * as utils from 'src/utils';
 import cookie from 'src/cookie';
